+++ src/DotPulsar/Internal/PartitionedProducer.cs
@@ -3,9 +3,9 @@
     using Abstractions;
     using DotPulsar.Abstractions;
     using Events;
+    using Exceptions;
     using System;
     using System.Collections.Concurrent;
-    using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
 
@@ -13,13 +13,14 @@
     {
         private readonly Guid _correlationId;
         private readonly IRegisterEvent _eventRegister;
+        private readonly IExecute _executor;
         private readonly IStateChanged<ProducerState> _state;
         private readonly PulsarClient _pulsarClient;
         private readonly ProducerOptions<TMessage> _options;
         private readonly ConcurrentDictionary<int, IProducer<TMessage>> _producers;
         private readonly IMessageRouter _messageRouter;
         private readonly CancellationTokenSource _cts = new();
-        private readonly int _producersCount;
+        private int _producersCount;
         private int _isDisposed;
         public Uri ServiceUrl { get; }
         public string Topic { get; }
@@ -29,8 +30,8 @@
             Uri serviceUrl,
             string topic,
             IRegisterEvent registerEvent,
+            IExecute executor,
             IStateChanged<ProducerState> state,
-            uint partitionsCount,
             ProducerOptions<TMessage> options,
             PulsarClient pulsarClient
         )
@@ -39,24 +40,40 @@
             ServiceUrl = serviceUrl;
             Topic = topic;
             _eventRegister = registerEvent;
+            _executor = executor;
             _state = state;
             _isDisposed = 0;
             _options = options;
             _pulsarClient = pulsarClient;
-            _producersCount = (int) partitionsCount;
             _messageRouter = options.MessageRouter;
 
-            _producers = new ConcurrentDictionary<int, IProducer<TMessage>>(1, _producersCount);
-            CreateSubProducers(0, _producersCount);
+            _producers = new ConcurrentDictionary<int, IProducer<TMessage>>(1, 31);
+
+            UpdatePartitions(_cts.Token);
         }
 
         private void CreateSubProducers(int startIndex, int count)
         {
-            for (var i = 0; i < count; ++i)
+            if (count == 0)
             {
-                var producer = _pulsarClient.NewProducer(Topic, _options, (uint) (i + startIndex), _correlationId);
-                _producers[i + startIndex] = producer;
+                var producer = _pulsarClient.NewSubProducer(Topic, _options, _executor, _correlationId);
+                _producers[0] = producer;
+                return;
             }
+
+            for (var i = startIndex; i < count; ++i)
+            {
+                var producer = _pulsarClient.NewSubProducer(Topic, _options, _executor, _correlationId, (uint) i);
+                _producers[i] = producer;
+            }
+        }
+
+        private async void UpdatePartitions(CancellationToken cancellationToken)
+        {
+            var partitionsCount = (int) await _pulsarClient.GetNumberOfPartitions(Topic, cancellationToken).ConfigureAwait(false);
+            _eventRegister.Register(new UpdatePartitions(_correlationId, (uint)partitionsCount));
+            CreateSubProducers(_producers.Count, partitionsCount);
+            _producersCount = partitionsCount;
         }
 
         public bool IsFinalState()
@@ -87,10 +104,19 @@
             _eventRegister.Register(new ProducerDisposed(_correlationId));
         }
 
+        private int ChoosePartitions(MessageMetadata? metadata)
+        {
+            if (_producers.IsEmpty)
+            {
+                throw new LookupNotReadyException();
+            }
+            return _producersCount == 0 ? 0 : _messageRouter.ChoosePartition(metadata, _producersCount);
+        }
+
         public async ValueTask<MessageId> Send(TMessage message, CancellationToken cancellationToken = default)
-            => await _producers[_messageRouter.ChoosePartition(null, _producersCount)].Send(message, cancellationToken).ConfigureAwait(false);
+            => await _executor.Execute(() => _producers[ChoosePartitions(null)].Send(message, cancellationToken), cancellationToken).ConfigureAwait(false);
 
         public async ValueTask<MessageId> Send(MessageMetadata metadata, TMessage message, CancellationToken cancellationToken = default)
-            => await _producers[_messageRouter.ChoosePartition(metadata, _producersCount)].Send(message, cancellationToken).ConfigureAwait(false);
+            => await _executor.Execute(() => _producers[ChoosePartitions(metadata)].Send(message, cancellationToken), cancellationToken).ConfigureAwait(false);
     }
 }
